Utilities cleanup (#3677)
