+++ package.json
@@ -33,7 +33,7 @@
     "prettier": "^2.5.1",
     "prettier-plugin-prisma": "3.8.0",
     "pretty-quick": "3.1.3",
-    "turbo": "1.1.5",
+    "turbo": "1.4.2",
     "vitest": "0.8.2",
     "wait-on": "6.0.1"
   },
