+++ packages/blitz/src/index-browser.tsx
@@ -79,8 +79,7 @@ if (typeof window !== "undefined" && process.env.NODE_ENV === "development") {
 }
 
 export * from "./utils"
-export * from "./ts-utils"
 export * from "./types"
 export * from "./errors"
-export * from "./zod-utils"
-export * from "./prisma-utils"
+export * from "./utils/zod"
+export * from "./utils/prisma"
