+++ packages/blitz/src/index-server.ts
@@ -1,20 +1,14 @@
 import "./global"
 import {IncomingMessage, ServerResponse} from "http"
 import {Ctx} from "./types"
-// import {findBlitzConfigDirectory} from "./cli/utils/routes-manifest"
-// import {readFileSync} from "fs-extra"
 
 export * from "./index-browser"
 export * from "./types"
-export * from "./prisma-utils"
-export * from "./run-prisma"
+export * from "./utils/run-prisma"
 export * from "./middleware"
 export * from "./paginate"
 export {baseLogger, newLine, log} from "./logging"
 export {startWatcher, stopWatcher} from "./cli/utils/routes-manifest"
-// const blitzConfig = findBlitzConfigDirectory() as string
-// const file = readFileSync(blitzConfig)
-// export const Routes = eval(file.toString())
 
 export interface MiddlewareResponse<C extends Ctx = Ctx> extends ServerResponse {
   blitzCtx: C
